% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_list.R
\docType{data}
\name{currency_list}
\alias{currency_list}
\title{Currency names and codes and names of 200+ countries}
\format{A data frame with 249 rows and 7 variables:
\describe{
  \item{countrycode_2}{2-letter country code}
  \item{countrycode_3}{3-letter country code}
  \item{countrycode_num}{country numeric code}
  \item{country_name}{common counry name}
  \item{country_fullname}{full country name}
  \item{currency_code}{currency ISO code}
  \item{currency_name}{full currency name}


}}
\source{
\code{countrycode} package \url{https://github.com/vincentarelbundock/countrycode/}
}
\usage{
currency_list
}
\description{
A dataset containing the currency name, codes and other country attributes of almost 250 coutries.
}
\keyword{datasets}
